+++ R/aux_hidden.R
@@ -9,14 +9,36 @@
 # 03. hidden_cmds     : Classical Multidimensional Scaling
 # 04. hidden_kmeanspp : k-means++ algorithm.
 # 05. hidden_tsne     : t-SNE visualization.
+# 06. hidden_nem      : Negative Eigenvalue Magnitude
+# 07. hidden_nef      : Negative Eigenfraction
+# 08. hidden_emds     : Euclified Multidimensional Scaling
 
 
 
+# 00. hidden_checker ------------------------------------------------------
+#' @keywords internal
+hidden_checker <- function(xobj){
+  if (inherits(xobj, "dist")){
+    return(as.matrix(xobj))
+  } else if (inherits(xobj, "matrix")){
+    check1 = (nrow(xobj)==ncol(xobj))
+    check2 = isSymmetric(xobj)
+    check3 = all(abs(diag(xobj))<.Machine$double.eps*10)
+    if (check1&&check2&&check3){
+      return(as.matrix(xobj))
+    } else {
+      stop("* hidden : matrix is not valid.")
+    }
+  } else {
+    stop("* hidden : input is not valid.")
+  }
+}
 
 # 01. hidden_kmedoids -----------------------------------------------------
 #' @keywords internal
 hidden_kmedoids <- function(distobj, nclust=2){
-  myk = round(nclust)
+  distobj = stats::as.dist(hidden_checker(distobj))
+  myk     = round(nclust)
   return(cluster::pam(distobj, k = myk))
 }
 
@@ -69,13 +91,13 @@ hidden_bmds <- function(x, ndim=2, par.a=5, par.alpha=0.5, par.step=1, mc.iter=8
 hidden_cmds <- function(x, ndim=2){
   ##################################################3
   # Check Input and Transform
+  x    = hidden_checker(x)
   ndim = round(ndim)
   k  = as.integer(ndim)
-  x  = as.matrix(x)
   D2 = (x^2)          # now squared matrix
   n  = nrow(D2)
   if ((length(ndim)>1)||(ndim<1)||(ndim>=nrow(x))){
-    stop("* cmds - 'ndim' should be an integer in [1,ncol(data)). ")
+    stop("* cmds : 'ndim' should be an integer in [1,ncol(data)). ")
   }
   
   ##################################################3
@@ -102,7 +124,7 @@ hidden_cmds <- function(x, ndim=2){
 hidden_kmeanspp <- function(x, k=2){
   ##################################################3
   # Check Input and Transform
-  x  = as.matrix(x)
+  x  = hidden_checker(x)
   n  = nrow(x)
   K  = round(k)
   if (K >= n){
@@ -150,8 +172,8 @@ hidden_kmeanspp <- function(x, k=2){
 hidden_tsne <- function(dx, ndim=2, ...){
   ##################################################
   # Pass to 'Rtsne'
+  dx     = hidden_checker(dx)
   k      = round(ndim)
-  dx     = as.matrix(dx)
   tmpout = Rtsne::Rtsne(dx, dims=k, ..., is_distance=TRUE)
   Y   = tmpout$Y
   DY  = as.matrix(stats::dist(Y))
@@ -163,3 +185,118 @@ hidden_tsne <- function(dx, ndim=2, ...){
   output$stress = compute_stress(dx, DY)
   return(output)
 }
+
+# 06. hidden_nem ----------------------------------------------------------
+#' @keywords internal
+hidden_nem <- function(xdiss){
+  ##################################################3
+  # Check Input and Transform
+  xx = hidden_checker(xdiss)
+  D2 = (xx^2)
+  n  = nrow(D2)
+
+  ##################################################3
+  # Computation
+  H = diag(n) - (1/n)*base::outer(rep(1,n),rep(1,n))
+  S = -0.5*(H%*%D2%*%H)
+  eigS = base::eigen(S, only.values = TRUE)
+  evals = eigS$values
+
+  ##################################################3
+  # Finalize
+  output = abs(min(evals))/max(evals)
+  return(output)
+}
+
+
+# 07. hidden_nef ----------------------------------------------------------
+#' @keywords internal
+hidden_nef <- function(xdiss){
+  ##################################################3
+  # Check Input and Transform
+  xx = hidden_checker(xdiss)
+  D2 = (xx^2)
+  n  = nrow(D2)
+  
+  ##################################################3
+  # Computation
+  H = diag(n) - (1/n)*base::outer(rep(1,n),rep(1,n))
+  S = -0.5*(H%*%D2%*%H)
+  eigS = base::eigen(S)
+  evals = eigS$values
+  
+  ##################################################3
+  # Finalize
+  output = sum(abs(evals[which(evals<0)]))/sum(abs(evals))
+  return(output)
+}
+
+# 08. hidden_emds ---------------------------------------------------------
+#' Euclified Multidimensional Scaling
+#' 
+#' strategy 1 : transitive closure of the triangle inequality (labdsv)
+#' strategy 2 : Non-Euclidean or Non-metric Measures Can Be Informative; adding positive numbers to all off-diagonal entries
+#' 
+#' @keywords internal
+#' @noRd
+hidden_emds <- function(xdiss, ndim=2, method=c("closure","gram")){
+  ##################################################3
+  # Check Input and Transform
+  x    = hidden_checker(xdiss)
+  ndim = round(ndim)
+  k  = as.integer(ndim)
+  n  = nrow(x)
+  if ((length(ndim)>1)||(ndim<1)||(ndim>=nrow(x))){
+    stop("* emds : 'ndim' should be an integer in [1,nrow(x)). ")
+  }
+  
+  method = match.arg(method) 
+  mydim  = round(ndim)
+  
+  ##################################################3
+  # Branching
+  if (hidden_nef(x) < 100*.Machine$double.eps){ # if Euclidean, okay
+    output = hidden_cmds(x, ndim=mydim)
+  } else { # if not Euclidean
+    if (method=="closure"){ # strategy 1 : transitive closure of the triangle inequality
+      xnew   = as.matrix(labdsv::euclidify(stats::as.dist(x))) # well it seems to work well..
+      output = hidden_cmds(xnew, ndim = mydim)
+    } else {                # strategy 2 : add positive numbers to all off-diagonal entries
+      gamma0 = emds_gamma0(x)
+      ggrid  = seq(from=min(0.001, gamma0/1000), to=(gamma0*0.999), length.out=20) # just try 20 cases
+      vgrid  = rep(0,20)
+      for (i in 1:20){
+        xtmp = x + ggrid[i]
+        diag(xtmp) = rep(0,nrow(xtmp))
+        vgrid[i]   = hidden_nef(xtmp)
+      }
+      idopts = which.min(vgrid)
+      if (length(idopts)>1){ # if multiple, use the first one.
+        idopts = idopts[1]
+      }
+      optgamma   = ggrid[idopts]
+      xnew       = x + optgamma
+      diag(xnew) = rep(0,nrow(xnew))
+      output     = hidden_cmds(xnew, ndim = mydim)
+    } 
+  }
+  
+  ##################################################3
+  # Report 
+  return(output)
+}
+
+# # example -----------------------------------------------------------------
+# library(labdsv)
+# data(bryceveg) # returns a vegetation data.frame
+# dis.bc <- as.matrix(dsvdis(bryceveg,'bray/curtis')) # calculate a Bray/Curtis
+# 
+# emds = getFromNamespace("hidden_emds","maotai")
+# out.cmds <- cmds(dis.bc, ndim=2)$embed
+# out.emds1 <- emds(dis.bc, ndim=2, method="closure")$embed
+# out.emds2 <- emds(dis.bc, ndim=2, method="gram")$embed
+# 
+# par(mfrow=c(3,1))
+# plot(out.cmds, main="cmds")
+# plot(out.emds1, main="emds::closure")
+# plot(out.emds2, main="emds::gram")